Version history cleanup
